fix: MULTIARCH-4654 needed a slightly different path and approach
