+++ internal/pkg/daemon/bpfrecorder/bpfrecorder_util_default.go
@@ -1,5 +1,5 @@
-//go:build amd64 && linux
-// +build amd64,linux
+//go:build !ppc64le && !s390x
+// +build !ppc64le,!s390x
 
 /*
 Copyright 2024 The Kubernetes Authors.
@@ -19,7 +19,10 @@ limitations under the License.
 
 package bpfrecorder
 
-import "syscall"
+import (
+	"strings"
+	"syscall"
+)
 
 // UnameMachineToString converts uname.Machine to a string for amd64.
 func UnameMachineToString(uname *syscall.Utsname) string {
@@ -29,4 +32,20 @@ func UnameMachineToString(uname *syscall.Utsname) string {
 // UnameReleaseToString converts uname.Release to a string for amd64.
 func UnameReleaseToString(uname *syscall.Utsname) string {
 	return toStringInt8(uname.Release)
+}
+
+func toStringInt8(array [65]int8) string {
+	var buf [65]byte
+	for i, b := range array {
+		buf[i] = byte(b)
+	}
+	return toStringByte(buf[:])
+}
+
+func toStringByte(array []byte) string {
+	str := string(array)
+	if i := strings.Index(str, "\x00"); i != -1 {
+		str = str[:i]
+	}
+	return str
 }
