+++ internal/pkg/daemon/bpfrecorder/bpfrecorder_util_linux_ppc64le.go
